Performance improvement for StringTokenizer.lookAhead(int)
Reverted bad optimization attempt via exception-driven branching introduced in 2a8aa8a162713361ef862afe0515de90392dfbe4.

Rationale:
Profiling shows a huge amount of CPU time wasted in this method in the constructor of the ArrayIndexOutOfBoundsException, which is only used for local branch control and is otherwise unnecessary. Creating the exception involves generating a stack trace; even if the probability of hitting an out-of-bounds lookAhead is low, doing this a few times is much more costly than what we save by avoiding the addition and comparison combo in the in-bounds cases. Note that exception handling introduces branching in the method anyway.